[Refactor] 생성 클래스 패키지 util로 분리
